tests/transaction: Modularize transaction implementations
