+++ tests/src/perun/test/transaction/mod.rs
@@ -1,0 +1,13 @@
+mod open;
+pub use open::*;
+
+mod fund;
+pub use fund::*;
+
+mod abort;
+pub use abort::*;
+
+mod close;
+pub use close::*;
+
+mod common;
